Yair/deprecate declared classes (#27)
* deprecate declared classes

* Move ContractClass into deprecated_contract_class

* Fix documentation
